requirements.txt
lxml
random_user_agent
httpx
aiosqlite
setuptools
typing
python-dotenv